+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQNQ1F78DJZNEY65VZ14X5JT"
+        "01JQPKBGKQ0YN3BN71Z0P51P3C"
     )
     r"""The id of the resource"""