+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQNQ1F7902J5G1235AJ2Y6QN"
+        "01JQPKBGKRCYSQVPTTD4E2CD21"
     )
     r"""The id of the resource"""